refactor to move 'neighbors' field out of Tile class
Creates a new field on Hexagon that maps tile IDs to sets of other tile IDs, as an alternative to storing the set of neighbors in each Tile object itself.
The set of Tiles in Hexagon is ordered, so a tile's index into that set is the key into the neighbors map.